+++ bar/bar/ringbuffers.h
@@ -186,7 +186,7 @@ typedef char(*RingBufferElementIterateFunction)(void *userData, void *data);
 #ifdef NDEBUG
 bool RingBuffer_init(RingBuffer *ringBuffer, uint elementSize, ulong size);
 #else /* not NDEBUG */
-bool __RingBuffer_init(const char *__fileName__, ulong __lineNb__, RingBuffer *ringBuffer, uint elementSize, ulong size);
+bool __RingBuffer_init(const char *__fileName__, uint __lineNb__, RingBuffer *ringBuffer, uint elementSize, ulong size);
 #endif /* NDEBUG */
 
 /***********************************************************************\
@@ -204,7 +204,7 @@ bool __RingBuffer_init(const char *__fileName__, ulong __lineNb__, RingBuffer *r
 #ifdef NDEBUG
 void RingBuffer_done(RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
 #else /* not NDEBUG */
-void __RingBuffer_done(const char *__fileName__, ulong __lineNb__, RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
+void __RingBuffer_done(const char *__fileName__, uint __lineNb__, RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
 #endif /* NDEBUG */
 
 /***********************************************************************\
@@ -220,7 +220,7 @@ void __RingBuffer_done(const char *__fileName__, ulong __lineNb__, RingBuffer *r
 #ifdef NDEBUG
 RingBuffer* RingBuffer_new(uint elementSize, ulong size);
 #else /* not NDEBUG */
-RingBuffer* __RingBuffer_new(const char *__fileName__, ulong __lineNb__, uint elementSize, ulong size);
+RingBuffer* __RingBuffer_new(const char *__fileName__, uint __lineNb__, uint elementSize, ulong size);
 #endif /* NDEBUG */
 
 /***********************************************************************\
@@ -238,7 +238,7 @@ RingBuffer* __RingBuffer_new(const char *__fileName__, ulong __lineNb__, uint el
 #ifdef NDEBUG
 void RingBuffer_delete(RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
 #else /* not NDEBUG */
-void __RingBuffer_delete(const char *__fileName__, ulong __lineNb__, RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
+void __RingBuffer_delete(const char *__fileName__, uint __lineNb__, RingBuffer *ringBuffer, RingBufferElementFreeFunction ringBufferElementFreeFunction, void *ringBufferElementFreeUserData);
 #endif /* NDEBUG */
 
 /***********************************************************************\
